+++ nion/swift/test/HistogramPanel_test.py
@@ -130,7 +130,7 @@ class TestHistogramPanelClass(unittest.TestCase):
             target_display_item_stream.remove_ref()
 
     def test_cursor_histogram_of_empty_data_displays_without_exception(self):
-        self.data_item.set_xdata(DataAndMetadata.DataAndMetadata(lambda: None, ((0, 0), numpy.float)))
+        self.data_item.set_xdata(DataAndMetadata.DataAndMetadata(lambda: None, ((0, 0), float)))
         self.histogram_canvas_item.mouse_position_changed(80, 58, 0)
 
     def test_histogram_statistics_with_zero_array(self):
